build fix

+++ wallet/wallet2.sol
@@ -46,7 +46,7 @@ contract multiowned {
     }
     // Revokes a prior confirmation of the given operation
     function revoke(bytes32 _operation) external {
-        uint ownerIndex = m_ownerIndex[msg.sender];
+        uint ownerIndex = m_ownerIndex[uint(msg.sender)];
         // make sure they're an owner
         if (ownerIndex == 0) return;
         uint ownerIndexBit = 2**ownerIndex;
